compiler_invocation banned in system tests.
1. YAML test specs are too big a surface to defend against malicious
code. People running Yuno normally can always check compiler_invocation
in their settings file, but that becomes less practical with dozens of
test cases.

I don't see a simple way to keep a system meant to run a user-specified
command from running a command which happens to be nasty, so for now
that system is disabled.

2. Removed compiler_invocation from some tests. They didn't need it
anyway.